requirements.txt
# please make sure you have already a pytorch install that is cuda enabled!
av
einops
flashy>=0.0.1
hydra-core>=1.1
hydra_colorlog
julius
num2words
numpy
sentencepiece
spacy==3.5.2
# torch>=2.0.0
# torchaudio>=2.0.0
huggingface_hub
tqdm
transformers>=4.31.0  # need Encodec there.
xformers
demucs
librosa
gradio
torchmetrics
encodec
pytaglib
python-dotenv
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
gdown